requirements.txt
langchain
flask
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
chroma
pypdf
flask-cors